+++ packages/artifact/__tests__/util.test.ts
@@ -1,10 +1,10 @@
 import * as fs from 'fs'
 import * as io from '../../io/src/io'
 import * as path from 'path'
-import * as utils from '../src/internal-utils'
+import * as utils from '../src/internal/utils'
 import * as core from '@actions/core'
 import {HttpCodes} from '@actions/http-client'
-import {getRuntimeUrl, getWorkFlowRunId} from '../src/internal-config-variables'
+import {getRuntimeUrl, getWorkFlowRunId} from '../src/internal/config-variables'
 
 jest.mock('../src/internal-config-variables')
