+++ packages/artifact/src/__mocks__/internal/config-variables.ts
